PalindroneFinder
Methods to test whether a string is a palindrone and to find all
palindrones in a given string.

The method for finding all palindrones is designed to minimize
comparisons by only looking at strings of length i or m - i
where i is the current index and m is the length of the string.